refactor: 순수계산함수 디렉토리 변경
